Added testing framework for bdd - ginkgo
